fix(position): CtCatch has no modifiers, they are in CatchVariable (#2156)
The modifiers were on two places before
1) CtCatch - wrong
2) CtCactchVariable - OK

So this PR removes them from CtCatch

Note: CtCatch would need a bettet source position structure. I actually I used BodyHolderSourcePosition, but it is may be not fitting well here. WDYT? Can we live with that or we make new structure for that?