fix(position): Invalid column of the element, positioned before the first line separator (#2409)
Hi! I've found that spoon gives incorrect column number for the element, positioned before the first line separator in file.

Consider the following file (TestSimpleClass.java):
```
public class TestSimpleClass {
    public int x;
    public int y;
    public int z;
}
```
If you try to get column from postion of this class, you will get -69, which is incorrect.
So, this PR is intended to fix that.